refactor: useScroll 훅 분리
